........
  r85679 | georg.brandl | 2010-10-18 07:06:18 +0200 (Mo, 18 Okt 2010) | 1 line

  Fix compiler warnings about formatting pid_t as an int, by always casting to long.
........

+++ Modules/_multiprocessing/semaphore.c
@@ -433,7 +433,7 @@ semlock_new(PyTypeObject *type, PyObject *args, PyObject *kwds)
         return NULL;
     }
 
-    PyOS_snprintf(buffer, sizeof(buffer), "/mp%d-%d", getpid(), counter++);
+    PyOS_snprintf(buffer, sizeof(buffer), "/mp%ld-%d", (long)getpid(), counter++);
 
     SEM_CLEAR_ERROR();
     handle = SEM_CREATE(buffer, value, maxvalue);
